+++ python/ray/data/_internal/execution/interfaces/task_context.py
@@ -2,7 +2,7 @@ import threading
 from dataclasses import dataclass, field
 from typing import TYPE_CHECKING, Any, Dict, Optional
 
-from ray.data._internal.progress_bar import ProgressBar
+from ray.data._internal.progress.base_progress import BaseProgressBar
 
 if TYPE_CHECKING:
     from ray.data._internal.execution.operators.map_transformer import MapTransformer
@@ -25,7 +25,7 @@ class TaskContext:
     # The dictionary of sub progress bar to update. The key is name of sub progress
     # bar. Note this is only used on driver side.
     # TODO(chengsu): clean it up from TaskContext with new optimizer framework.
-    sub_progress_bar_dict: Optional[Dict[str, ProgressBar]] = None
+    sub_progress_bar_dict: Optional[Dict[str, BaseProgressBar]] = None
 
     # NOTE(hchen): `upstream_map_transformer` and `upstream_map_ray_remote_args`
     # are only used for `RandomShuffle`. DO NOT use them for other operators.
